Add and use terminator line logic in client library
There are two ways the client library parses server responses: expect a
fixed number of lines and return when they have been received, or wait
for an interval of time then return anything received in that interval.

In the future, we may add a terminator such as a blank line to
variable-length responses, so the logic in this commit handles these
terminators in a forward-compatible way.

By treating '0 disabled' as a terminator line, we can also use this
logic now to return early if the dump feature of the server is disabled.