Implement simple, optional PID file
glockd calls flock() on the PID file, which was not done here because
flock is not available in the Node.js standard libraries.  There are
libraries available that have it, but the drawback of using them is that
they require a compilation step.  `npm install` will take care of
compilation steps seamlessly if the environment is set up correctly, but
this is difficult to do e.g. on Windows.