Ignore blank lines from clients
Due to the way the split() stream parses lines, the server code receives
a blank line when a client disconnects.  There may be a cleaner way to
prevent this from happening, but since there's never a situation where
we need to do something with a blank line, just ignore them.

At this point the tests pass with all server features enabled - probably
not with features disabled just yet, though.